+++ docs/core_docs/docs/integrations/callbacks/datadog_tracer.mdx
@@ -1,10 +1,10 @@
 ---
-sidebar_class_name: node-only
+sidebar_class_name: beta
 ---
 
 import CodeBlock from "@theme/CodeBlock";
 
-# Datadog LLM Observability Tracer Callback
+# Datadog LLM Observability
 
 :::warning
 LLM Observability is in public beta, and its API is subject to change.
@@ -12,11 +12,7 @@ LLM Observability is in public beta, and its API is subject to change.
 
 With [Datadog LLM Observability](https://docs.datadoghq.com/llm_observability/), you can monitor, troubleshoot, and evaluate your LLM-powered applications, such as chatbots. You can investigate the root cause of issues, monitor operational performance, and evaluate the quality, privacy, and safety of your LLM applications.
 
-This is a community implementation, and it is not officially supported by Datadog. It is based on the [Datadog LLM Observability API](https://docs.datadoghq.com/llm_observability/api).
-
-:::tip Compatibility
-Only available on Node.js.
-:::
+This is an experimental community implementation, and it is not officially supported by Datadog. It is based on the [Datadog LLM Observability API](https://docs.datadoghq.com/llm_observability/api).
 
 ## Setup
