+++ libs/langchain-community/.gitignore
@@ -610,10 +610,6 @@ callbacks/handlers/upstash_ratelimit.cjs
 callbacks/handlers/upstash_ratelimit.js
 callbacks/handlers/upstash_ratelimit.d.ts
 callbacks/handlers/upstash_ratelimit.d.cts
-callbacks/tracers/datadog.cjs
-callbacks/tracers/datadog.js
-callbacks/tracers/datadog.d.ts
-callbacks/tracers/datadog.d.cts
 retrievers/amazon_kendra.cjs
 retrievers/amazon_kendra.js
 retrievers/amazon_kendra.d.ts
@@ -1014,6 +1010,10 @@ utils/cassandra.cjs
 utils/cassandra.js
 utils/cassandra.d.ts
 utils/cassandra.d.cts
+experimental/callbacks/handlers/datadog.cjs
+experimental/callbacks/handlers/datadog.js
+experimental/callbacks/handlers/datadog.d.ts
+experimental/callbacks/handlers/datadog.d.cts
 experimental/graph_transformers/llm.cjs
 experimental/graph_transformers/llm.js
 experimental/graph_transformers/llm.d.ts
