+++ libs/langchain-community/package.json
@@ -2077,15 +2077,6 @@
       "import": "./callbacks/handlers/upstash_ratelimit.js",
       "require": "./callbacks/handlers/upstash_ratelimit.cjs"
     },
-    "./callbacks/tracers/datadog": {
-      "types": {
-        "import": "./callbacks/tracers/datadog.d.ts",
-        "require": "./callbacks/tracers/datadog.d.cts",
-        "default": "./callbacks/tracers/datadog.d.ts"
-      },
-      "import": "./callbacks/tracers/datadog.js",
-      "require": "./callbacks/tracers/datadog.cjs"
-    },
     "./retrievers/amazon_kendra": {
       "types": {
         "import": "./retrievers/amazon_kendra.d.ts",
@@ -2986,6 +2977,15 @@
       "import": "./utils/cassandra.js",
       "require": "./utils/cassandra.cjs"
     },
+    "./experimental/callbacks/handlers/datadog": {
+      "types": {
+        "import": "./experimental/callbacks/handlers/datadog.d.ts",
+        "require": "./experimental/callbacks/handlers/datadog.d.cts",
+        "default": "./experimental/callbacks/handlers/datadog.d.ts"
+      },
+      "import": "./experimental/callbacks/handlers/datadog.js",
+      "require": "./experimental/callbacks/handlers/datadog.cjs"
+    },
     "./experimental/graph_transformers/llm": {
       "types": {
         "import": "./experimental/graph_transformers/llm.d.ts",
@@ -3647,10 +3647,6 @@
     "callbacks/handlers/upstash_ratelimit.js",
     "callbacks/handlers/upstash_ratelimit.d.ts",
     "callbacks/handlers/upstash_ratelimit.d.cts",
-    "callbacks/tracers/datadog.cjs",
-    "callbacks/tracers/datadog.js",
-    "callbacks/tracers/datadog.d.ts",
-    "callbacks/tracers/datadog.d.cts",
     "retrievers/amazon_kendra.cjs",
     "retrievers/amazon_kendra.js",
     "retrievers/amazon_kendra.d.ts",
@@ -4051,6 +4047,10 @@
     "utils/cassandra.js",
     "utils/cassandra.d.ts",
     "utils/cassandra.d.cts",
+    "experimental/callbacks/handlers/datadog.cjs",
+    "experimental/callbacks/handlers/datadog.js",
+    "experimental/callbacks/handlers/datadog.d.ts",
+    "experimental/callbacks/handlers/datadog.d.cts",
     "experimental/graph_transformers/llm.cjs",
     "experimental/graph_transformers/llm.js",
     "experimental/graph_transformers/llm.d.ts",
